src/app/requirements.txt
flask==2.1
requests==2.23.0
flask-http-middleware==0.3.1
prometheus-client==0.17.0
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-flask==0.38b0
opentelemetry-instrumentation-requests==0.38b0
opentelemetry-instrumentation-logging==0.38b0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-exporter-prometheus==0.38b0
